Refactor experiment/image-bumper and robots/pr-creator for reuse.
