+++ experiment/image-bumper/main.go
@@ -17,171 +17,13 @@ limitations under the License.
 package main
 
 import (
-	"encoding/json"
 	"flag"
-	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
 	"regexp"
-	"strconv"
 	"strings"
-	"time"
+
+	"k8s.io/test-infra/experiment/image-bumper/bumper"
 )
-
-var (
-	imageRegexp = regexp.MustCompile(`\b((?:[a-z0-9]+\.)?gcr\.io)/([a-z][a-z0-9-]{5,29}/[a-zA-Z0-9][a-zA-Z0-9_.-]+):([a-zA-Z0-9_.-]+)\b`)
-	tagRegexp   = regexp.MustCompile(`(v?\d{8}-(?:v\d(?:[.-]\d+)*-g)?[0-9a-f]{6,10}|latest)(-.+)?`)
-	tagCache    = map[string]string{}
-	httpClient  = http.Client{Timeout: 1 * time.Minute}
-)
-
-const (
-	imageHostPart  = 1
-	imageImagePart = 2
-	imageTagPart   = 3
-	tagVersionPart = 1
-	tagExtraPart   = 2
-)
-
-type manifest map[string]struct {
-	TimeCreatedMs string   `json:"timeCreatedMs"`
-	Tags          []string `json:"tag"`
-}
-
-func findLatestTag(imageHost, imageName, currentTag string) (string, error) {
-	k := imageHost + "/" + imageName + ":" + currentTag
-	if result, ok := tagCache[k]; ok {
-		return result, nil
-	}
-
-	currentTagParts := tagRegexp.FindStringSubmatch(currentTag)
-	if currentTagParts == nil {
-		return "", fmt.Errorf("couldn't figure out the current tag in %q", currentTag)
-	}
-	if currentTagParts[tagVersionPart] == "latest" {
-		return currentTag, nil
-	}
-
-	resp, err := httpClient.Get("https://" + imageHost + "/v2/" + imageName + "/tags/list")
-	if err != nil {
-		return "", fmt.Errorf("couldn't fetch tag list: %v", err)
-	}
-	defer resp.Body.Close()
-
-	result := struct {
-		Manifest manifest `json:"manifest"`
-	}{}
-
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		return "", fmt.Errorf("couldn't parse tag information from registry: %v", err)
-	}
-
-	latestTag, err := pickBestTag(currentTagParts, result.Manifest)
-	if err != nil {
-		return "", err
-	}
-
-	tagCache[k] = latestTag
-
-	return latestTag, nil
-}
-
-func pickBestTag(currentTagParts []string, manifest manifest) (string, error) {
-	// The approach is to find the most recently created image that has the same suffix as the
-	// current tag. However, if we find one called "latest" (with appropriate suffix), we assume
-	// that's the latest regardless of when it was created.
-	var latestTime int64
-	latestTag := ""
-	for _, v := range manifest {
-		bestVariant := ""
-		override := false
-		for _, t := range v.Tags {
-			parts := tagRegexp.FindStringSubmatch(t)
-			if parts == nil {
-				continue
-			}
-			if parts[tagExtraPart] != currentTagParts[tagExtraPart] {
-				continue
-			}
-			if parts[tagVersionPart] == "latest" {
-				override = true
-				continue
-			}
-			if bestVariant == "" || len(t) < len(bestVariant) {
-				bestVariant = t
-			}
-		}
-		if bestVariant == "" {
-			continue
-		}
-		t, err := strconv.ParseInt(v.TimeCreatedMs, 10, 64)
-		if err != nil {
-			return "", fmt.Errorf("couldn't parse timestamp %q: %v", v.TimeCreatedMs, err)
-		}
-		if override || t > latestTime {
-			latestTime = t
-			latestTag = bestVariant
-			if override {
-				break
-			}
-		}
-	}
-
-	if latestTag == "" {
-		return "", fmt.Errorf("failed to find a good tag")
-	}
-
-	return latestTag, nil
-}
-
-func updateAllTags(tagPicker func(host, image, tag string) (string, error), content []byte, imageFilter *regexp.Regexp) []byte {
-	indexes := imageRegexp.FindAllSubmatchIndex(content, -1)
-	// Not finding any images is not an error.
-	if indexes == nil {
-		return content
-	}
-
-	newContent := make([]byte, 0, len(content))
-	lastIndex := 0
-	for _, m := range indexes {
-		newContent = append(newContent, content[lastIndex:m[imageTagPart*2]]...)
-		host := string(content[m[imageHostPart*2]:m[imageHostPart*2+1]])
-		image := string(content[m[imageImagePart*2]:m[imageImagePart*2+1]])
-		tag := string(content[m[imageTagPart*2]:m[imageTagPart*2+1]])
-		lastIndex = m[1]
-
-		if tag == "" || (imageFilter != nil && !imageFilter.MatchString(host+"/"+image+":"+tag)) {
-			newContent = append(newContent, content[m[imageTagPart*2]:m[1]]...)
-			continue
-		}
-
-		latest, err := tagPicker(host, image, tag)
-		if err != nil {
-			log.Printf("Failed to update %s/%s:%s: %v.\n", host, image, tag, err)
-			newContent = append(newContent, content[m[imageTagPart*2]:m[1]]...)
-			continue
-		}
-		newContent = append(newContent, []byte(latest)...)
-	}
-	newContent = append(newContent, content[lastIndex:]...)
-
-	return newContent
-}
-
-func updateFile(path string, imageFilter *regexp.Regexp) error {
-	content, err := ioutil.ReadFile(path)
-	if err != nil {
-		return fmt.Errorf("failed to read %s: %v", path, err)
-	}
-
-	newContent := updateAllTags(findLatestTag, content, imageFilter)
-
-	if err := ioutil.WriteFile(path, newContent, 0644); err != nil {
-		return fmt.Errorf("failed to write %s: %v", path, err)
-	}
-	return nil
-}
 
 type options struct {
 	imageRegex string
@@ -207,12 +49,12 @@ func main() {
 		}
 	}
 	for _, f := range o.files {
-		if err := updateFile(f, imageRegex); err != nil {
+		if err := bumper.UpdateFile(f, imageRegex); err != nil {
 			log.Printf("Failed to update %s: %v", f, err)
 		}
 	}
 	log.Println("Done.")
-	for before, after := range tagCache {
+	for before, after := range bumper.GetReplacements() {
 		if strings.Split(before, ":")[1] == after {
 			continue
 		}
